pyproject.toml
[project]
name = "unit-converter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.3.0",
    "pint>=0.24.4",
    "streamlit",
]
